Refactor function imports in leaderboard and api_or_sql files to use new namespace for improved organization and clarity
